refactor: integrate RdbmsExporterConfig as builder into RdbmsExporter #35304
